Limit "Daily initiator" builder to build master
This Builder's sole responsibility is determining the latest revision of
WPT once every 12 hours. Configuring it to run on the generic workers
does not improve performance, and it requires that all workers maintain
duplicate clones of the WPT git repository.

Configure the builder to only execute on the build master.