Capture formatted logging information via Buildbot
Previously, information about the process of verifying the expected
results from the WPT CLI was emitted in the same stream as the WPT CLI
process's output. This made diagnosing errors more difficult because the
latter stream contains a far greater amount of data.

Configure the WPT CLI to emit structured logging data into a dedicated
file, and configure Buildbot to collect and display the contents of that
file.