Collect results from stable Safari locally
This project was previously extended with support for collecting results
from Apple Safari Technology Preview via a worker running on a macOS
system. Extend the system to also support the stable release of Apple
Safari.

Previously, a single build step installed Safari Technology Preview and
enabled remote automation. Because the stable release of Safari is
bundled with the operating system, it cannot be installed. It does
require explicit enabling of remote automation, however. Refactor the
two operations into distinct build steps in order to support both
browsers.

Safari Technology Preview enables the "Develop" UI menu by default, so
the automation script could interact with it immediately. The menu is
not present in the default configuration of Safari. Add logic to ensure
the menu is available.

This project already included logic to determine whether automation had
successfully been enabled in Safari. That logic assumed the browser
implemented the W3C WebDriver protocol, but the stable release of Safari
does not. Extend the script to support both

This change is not supported by an increase in infrastructure. Decrease
the rate of collection in order to avoid over-utilization.